Refactor subtensor transaction extension
